user_service/requirements.txt
pyodbc
redis
PyJWT